[#1600] Move content hash handling into ipfs package
