+++ core/completion.go
@@ -463,7 +463,7 @@ func (n *OpenBazaarNode) ValidateAndSaveRating(contract *pb.RicardianContract) (
 			continue
 		}
 
-		mh, err := EncodeMultihash([]byte(ratingJSON))
+		mh, err := ipfs.EncodeMultihash([]byte(ratingJSON))
 		if err != nil {
 			retErr = err
 			continue
