+++ core/order.go
@@ -8,7 +8,6 @@ import (
 	"errors"
 	"fmt"
 
-	"github.com/OpenBazaar/openbazaar-go/ipfs"
 	"github.com/golang/protobuf/ptypes/timestamp"
 
 	crypto "gx/ipfs/QmTW4SdgBWq9GjsBsHeUx8WuGxzhgzAf88UMH2w62PC8yK/go-libp2p-crypto"
@@ -22,6 +21,7 @@ import (
 	ipfspath "gx/ipfs/QmQAgv6Gaoe2tQpcabqwKXKChp2MZ7i3UXv9DqTTaxCaTR/go-path"
 
 	"github.com/OpenBazaar/jsonpb"
+	"github.com/OpenBazaar/openbazaar-go/ipfs"
 	"github.com/OpenBazaar/openbazaar-go/pb"
 	"github.com/OpenBazaar/openbazaar-go/repo"
 	"github.com/OpenBazaar/wallet-interface"
@@ -508,7 +508,7 @@ func (n *OpenBazaarNode) createContractWithOrder(data *PurchaseData) (*pb.Ricard
 		if err != nil {
 			return nil, err
 		}
-		listingID, err := EncodeCID(ser)
+		listingID, err := ipfs.EncodeCID(ser)
 		if err != nil {
 			return nil, err
 		}
@@ -829,7 +829,7 @@ func (n *OpenBazaarNode) CalcOrderID(order *pb.Order) (string, error) {
 	if err != nil {
 		return "", err
 	}
-	id, err := EncodeMultihash(ser)
+	id, err := ipfs.EncodeMultihash(ser)
 	if err != nil {
 		return "", err
 	}
@@ -912,7 +912,7 @@ func (n *OpenBazaarNode) CalculateOrderTotal(contract *pb.RicardianContract) (ui
 		// Subtract any coupons
 		for _, couponCode := range item.CouponCodes {
 			for _, vendorCoupon := range l.Coupons {
-				id, err := EncodeMultihash([]byte(couponCode))
+				id, err := ipfs.EncodeMultihash([]byte(couponCode))
 				if err != nil {
 					return 0, err
 				}
@@ -1273,7 +1273,7 @@ collectListings:
 		if err != nil {
 			return err
 		}
-		listingID, err := EncodeCID(ser)
+		listingID, err := ipfs.EncodeCID(ser)
 		if err != nil {
 			return err
 		}
@@ -1625,7 +1625,7 @@ func ParseContractForListing(hash string, contract *pb.RicardianContract) (*pb.L
 		if err != nil {
 			return nil, err
 		}
-		listingID, err := EncodeCID(ser)
+		listingID, err := ipfs.EncodeCID(ser)
 		if err != nil {
 			return nil, err
 		}
