+++ core/order_test.go
@@ -4,6 +4,7 @@ import (
 	"testing"
 
 	"github.com/OpenBazaar/openbazaar-go/core"
+	"github.com/OpenBazaar/openbazaar-go/ipfs"
 	"github.com/OpenBazaar/openbazaar-go/pb"
 	"github.com/OpenBazaar/openbazaar-go/test"
 	"github.com/golang/protobuf/proto"
@@ -47,7 +48,7 @@ func TestOpenBazaarNode_CalculateOrderTotal(t *testing.T) {
 	if err != nil {
 		t.Error(err)
 	}
-	listingID, err := core.EncodeCID(ser)
+	listingID, err := ipfs.EncodeCID(ser)
 	if err != nil {
 		t.Error(err)
 	}
@@ -118,7 +119,7 @@ func TestOpenBazaarNode_CalculateOrderTotal(t *testing.T) {
 	if err != nil {
 		t.Error(err)
 	}
-	listingID, err = core.EncodeCID(ser)
+	listingID, err = ipfs.EncodeCID(ser)
 	if err != nil {
 		t.Error(err)
 	}
@@ -142,7 +143,7 @@ func TestOpenBazaarNode_CalculateOrderTotal(t *testing.T) {
 	if err != nil {
 		t.Error(err)
 	}
-	listingID, err = core.EncodeCID(ser)
+	listingID, err = ipfs.EncodeCID(ser)
 	if err != nil {
 		t.Error(err)
 	}
@@ -156,7 +157,7 @@ func TestOpenBazaarNode_CalculateOrderTotal(t *testing.T) {
 	}
 
 	// Test with coupon percent discount
-	couponHash, err := core.EncodeMultihash([]byte("testcoupon"))
+	couponHash, err := ipfs.EncodeMultihash([]byte("testcoupon"))
 	if err != nil {
 		t.Error(err)
 	}
@@ -172,7 +173,7 @@ func TestOpenBazaarNode_CalculateOrderTotal(t *testing.T) {
 	if err != nil {
 		t.Error(err)
 	}
-	listingID, err = core.EncodeCID(ser)
+	listingID, err = ipfs.EncodeCID(ser)
 	if err != nil {
 		t.Error(err)
 	}
@@ -187,7 +188,7 @@ func TestOpenBazaarNode_CalculateOrderTotal(t *testing.T) {
 	}
 
 	// Test with coupon percent discount
-	couponHash, err = core.EncodeMultihash([]byte("testcoupon2"))
+	couponHash, err = ipfs.EncodeMultihash([]byte("testcoupon2"))
 	if err != nil {
 		t.Error(err)
 	}
@@ -203,7 +204,7 @@ func TestOpenBazaarNode_CalculateOrderTotal(t *testing.T) {
 	if err != nil {
 		t.Error(err)
 	}
-	listingID, err = core.EncodeCID(ser)
+	listingID, err = ipfs.EncodeCID(ser)
 	if err != nil {
 		t.Error(err)
 	}
@@ -230,7 +231,7 @@ func TestOpenBazaarNode_CalculateOrderTotal(t *testing.T) {
 	if err != nil {
 		t.Error(err)
 	}
-	listingID, err = core.EncodeCID(ser)
+	listingID, err = ipfs.EncodeCID(ser)
 	if err != nil {
 		t.Error(err)
 	}
@@ -256,7 +257,7 @@ func TestOpenBazaarNode_CalculateOrderTotal(t *testing.T) {
 	if err != nil {
 		t.Error(err)
 	}
-	listingID, err = core.EncodeCID(ser)
+	listingID, err = ipfs.EncodeCID(ser)
 	if err != nil {
 		t.Error(err)
 	}
@@ -277,7 +278,7 @@ func TestOpenBazaarNode_CalculateOrderTotal(t *testing.T) {
 	if err != nil {
 		t.Error(err)
 	}
-	listingID, err = core.EncodeCID(ser)
+	listingID, err = ipfs.EncodeCID(ser)
 	if err != nil {
 		t.Error(err)
 	}
@@ -323,7 +324,7 @@ func TestOpenBazaarNode_CalculateOrderTotal(t *testing.T) {
 	if err != nil {
 		t.Error(err)
 	}
-	listingID, err = core.EncodeCID(ser)
+	listingID, err = ipfs.EncodeCID(ser)
 	if err != nil {
 		t.Error(err)
 	}
