+++ ipfs/cid.go
@@ -1,0 +1,52 @@
+package ipfs
+
+import (
+	"crypto/sha256"
+	"errors"
+
+	ma "gx/ipfs/QmTZBfrPJmjWsCvHEtX5FE6KimVJhsJg5sBbqEFYf4UZtL/go-multiaddr"
+	"gx/ipfs/QmTbxNB1NwDesLmKTscr4udL2tVP7MaxvXnD1D9yX7g3PN/go-cid"
+	ps "gx/ipfs/QmaCTz9RkrU13bm9kMB54f7atgqM4qkjDZpRwRoJiWXEqs/go-libp2p-peerstore"
+	mh "gx/ipfs/QmerPMzPk1mJVowm8KgmoknWa4yCYvvugMPsgWmDNUvDLW/go-multihash"
+)
+
+// EncodeCID - Hash with SHA-256 and encode as a multihash
+func EncodeCID(b []byte) (*cid.Cid, error) {
+	multihash, err := EncodeMultihash(b)
+	if err != nil {
+		return nil, err
+	}
+	id := cid.NewCidV1(cid.Raw, *multihash)
+	return &id, err
+}
+
+// EncodeMultihash - sha256 encode
+func EncodeMultihash(b []byte) (*mh.Multihash, error) {
+	h := sha256.Sum256(b)
+	encoded, err := mh.Encode(h[:], mh.SHA2_256)
+	if err != nil {
+		return nil, err
+	}
+	multihash, err := mh.Cast(encoded)
+	if err != nil {
+		return nil, err
+	}
+	return &multihash, err
+}
+
+// ExtractIDFromPointer Certain pointers, such as moderators, contain a peerID. This function
+// will extract the ID from the underlying PeerInfo object.
+func ExtractIDFromPointer(pi ps.PeerInfo) (string, error) {
+	if len(pi.Addrs) == 0 {
+		return "", errors.New("PeerInfo object has no addresses")
+	}
+	addr := pi.Addrs[0]
+	if addr.Protocols()[0].Code != ma.P_IPFS {
+		return "", errors.New("IPFS protocol not found in address")
+	}
+	val, err := addr.ValueForProtocol(ma.P_IPFS)
+	if err != nil {
+		return "", err
+	}
+	return val, nil
+}
